+++ infuse/patch.py
@@ -41,7 +41,7 @@ class RequestBreaker:
 
     @staticmethod
     def namespace(service_name):
-        return f"{settings.MMT_ENV}:{service_name}"
+        return settings.INFUSE_REDIS_KEY_NAMESPACE_TEMPLATE.format(env=settings.MMT_ENV, service_name=service_name)
 
     async def wrapped_request(self, wrapped, instance, args, kwargs):
         skip_breaker = kwargs.pop("skip_breaker", False)
@@ -54,9 +54,9 @@ class RequestBreaker:
             try:
                 return await breaker.call(wrapped, *args, **kwargs)
             except CircuitBreakerError as e:
-
-                error_logger.critical(f"[INFUSE] [{self.namespace(instance.service_name)}] {e.args[0]}")
-                msg = settings.SERVICE_UNAVAILABLE_MESSAGE.format(instance.service_name)
+                service_name = kwargs.get('service_name', None) or self.namespace(instance.service_name)
+                error_logger.critical(f"[INFUSE] [{service_name}] {e.args[0]}")
+                msg = settings.SERVICE_UNAVAILABLE_MESSAGE.format(service_name)
 
                 exc = exceptions.ServiceUnavailable503Error(description=msg,
                                                             error_code=InfuseErrorCodes.service_unavailable,